Refactor variables
This commit refactors the variables used in these playbooks into a
group_vars directory.  The `keyname` variable is now mandatory
and the playbook will fail to execute if it is not set.  The `branch`
variable is set to default to `master` if it is not provided.  The
`prefix` variable is added to enable tailoring the created instance name
to the users needs.  It will default to a random number seeded by the
local hostname value.  This will make the prefix idempotent so that that
the user cannot accidentally create additional instances.