Sleep for 3 minutes to allow bootstrap to complete
This commit introduces a sleep step after the Phobos instance creation
has completed. Prior to this commit, the reliance on python would fail
because it had not been installed yet. This step allows time for the
bootstrap installation of python to complete before attempting to run
ansible commands against it.